split itsystem tests into responsibilities
